backend/requirements.txt
Flask==3.0.3
Flask-Mail==0.9.1
Flask-Cors==4.0.0
Flask-Limiter==3.8.0
gunicorn==23.0.0
python-dotenv==1.0.1
email-validator==2.2.0
